requirements.txt
flask
flask-cas-ng
flask-sqlalchemy
psycopg2
